# docs/swagger.yaml
definitions:
  entity.ComputerInfoModel:
    properties:
      cpu_info:
        items:
          $ref: '#/definitions/entity.CpuInfoModel'
        type: array
      disk_info:
        items:
          $ref: '#/definitions/entity.DiskInfoModel'
        type: array
      host_info:
        $ref: '#/definitions/entity.HostInfoModel'
      mem_info:
        $ref: '#/definitions/entity.MemInfoModel'
      net_info:
        items:
          $ref: '#/definitions/entity.NetInfoModel'
        type: array
      pool_activate_num:
        type: integer
      pool_job_num:
        type: integer
    type: object
  entity.CpuInfoModel:
    properties:
      cpu_core_num:
        type: string
      cpu_frequency:
        type: string
      cpu_name:
        type: string
      cpu_physical_id:
        type: string
      cpu_usage:
        type: string
      cpu_vendor_id:
        type: string
    type: object
  entity.DiskInfoModel:
    properties:
      disk_avi:
        type: string
      disk_name:
        type: string
      disk_size:
        type: string
      disk_usage:
        type: string
      disk_used:
        type: string
      fs_type:
        type: string
      opts:
        type: string
    type: object
  entity.HostInfoModel:
    properties:
      host_id:
        type: string
      host_kernel_arch:
        type: string
      host_kernel_version:
        type: string
      host_name:
        type: string
      host_os:
        type: string
      host_os_version:
        type: string
    type: object
  entity.MemInfoModel:
    properties:
      mem_avi:
        type: string
      mem_total:
        type: string
      mem_usage:
        type: string
      mem_used:
        type: string
    type: object
  entity.NetInfoModel:
    properties:
      bytes_received:
        type: string
      bytes_sent:
        type: string
      net_name:
        type: string
      packets_received:
        type: string
      packets_sent:
        type: string
    type: object
  entity.ResultModel:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  entity.ServerModel:
    properties:
      IP:
        type: string
      colony:
        type: string
      name:
        type: string
      namespace:
        type: string
      port:
        type: string
      time:
        type: string
    type: object
info:
  contact:
    email: 1539989223@qq.com
    name: CYCLEWW
    url: https://www.wzxaugenstern.online/#/
  description: 分布式记账系统调度中心
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.wzxaugenstern.online/#/
  title: Themis
  version: "1.0"
paths:
  /message/beat:
    put:
      consumes:
      - application/json
      description: 服务心跳重置倒计时
      parameters:
      - description: 服务实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回true或false
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 服务心跳
      tags:
      - 服务层
  /message/election:
    put:
      consumes:
      - application/json
      description: 由领导者调用的新一轮选举接口。
      parameters:
      - description: 领导者或非领导者服务实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回true或false
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 选举
      tags:
      - 服务层
  /message/getLeader:
    post:
      consumes:
      - application/json
      description: 由其他服务调用的获取当前领导者接口。
      parameters:
      - description: 非领导者实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回领导者服务信息
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  $ref: '#/definitions/entity.ServerModel'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取领导者
      tags:
      - 服务层
  /message/getServers:
    post:
      consumes:
      - application/json
      description: 由当前领导者调用的获取领导者所领导的服务列表。
      parameters:
      - description: 领导者实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回被领导者的切片数组
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ServerModel'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前被领导者服务列表
      tags:
      - 服务层
  /message/getServersNum:
    post:
      consumes:
      - application/json
      description: 由所有服务调用获取当前集群服务数量。
      parameters:
      - description: 服务实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回集群服务数量
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 获取当前集群服务数量
      tags:
      - 服务层
  /message/register:
    post:
      consumes:
      - application/json
      description: 服务注册进中心
      parameters:
      - description: 服务实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回true或false
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 服务注册
      tags:
      - 服务层
  /operator/cancelDeleteInstance:
    delete:
      consumes:
      - application/json
      description: 由管理者调用删除黑名单中的实例信息。
      parameters:
      - description: 从黑名单中清除的实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回true或false
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 删除黑名单中的实例信息
      tags:
      - 管理层
  /operator/deleteColony:
    delete:
      consumes:
      - application/json
      description: 由管理者调用删除地区集群实例并拉入黑名单。
      parameters:
      - description: 被删除的服务地区信息（用服务实例信息包装）
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回true或false
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 删除地区集群实例并拉入黑名单
      tags:
      - 管理层
  /operator/deleteInstance:
    delete:
      consumes:
      - application/json
      description: 由管理者调用删除服务实例并拉入黑名单。
      parameters:
      - description: 被删除的服务实例信息
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回true或false
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 删除服务实例并拉入黑名单
      tags:
      - 管理层
  /operator/getClusterLeader:
    get:
      consumes:
      - application/json
      description: 由管理员调用获取中心集群领导者
      produces:
      - application/json
      responses:
        "200":
          description: 返回中心集群领导者名称
          schema:
            $ref: '#/definitions/entity.ResultModel'
      security:
      - ApiKeyAuth: []
      summary: 获取中心集群领导者
      tags:
      - 管理层
  /operator/getClusterStatus:
    get:
      consumes:
      - application/json
      description: 由管理员调用获取中心当前身份状态
      produces:
      - application/json
      responses:
        "200":
          description: 返回集群状态
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取中心当前身份状态
      tags:
      - 管理层
  /operator/getDeleteInstance:
    get:
      consumes:
      - application/json
      description: 由管理者调用的获取当前全部黑名单服务实例
      produces:
      - application/json
      responses:
        "200":
          description: 返回黑名单中服务实例切片数组
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ServerModel'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取全部黑名单服务实例
      tags:
      - 管理层
  /operator/getInstances:
    get:
      consumes:
      - application/json
      description: 由管理者调用的获取当前所有服务信息。
      produces:
      - application/json
      responses:
        "200":
          description: 返回服务实例切片数组
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ServerModel'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取全部服务实例
      tags:
      - 管理层
    post:
      consumes:
      - application/json
      description: 由获取命名空间与集群条件下的服务实例服务实例。
      parameters:
      - description: 封装的条件参数
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/entity.ServerModel'
      produces:
      - application/json
      responses:
        "200":
          description: 返回服务实例切片数组
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ServerModel'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取指定条件下的服务实例服务实例
      tags:
      - 管理层
  /operator/getStatus:
    get:
      consumes:
      - application/json
      description: 由管理员调用获取当前中心线程数
      produces:
      - application/json
      responses:
        "200":
          description: 返回电脑状态
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResultModel'
            - properties:
                data:
                  $ref: '#/definitions/entity.ComputerInfoModel'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取服务状态
      tags:
      - 管理层
swagger: "2.0"